Move Parser to a higher level
